This is an EDA file
The shark tank India, my favorite show of this year 2022, had been telecast on sony live. I enjoyed every single pitch of the season. 
And was always wondering, how much amount have the sharks invested and which of them will make profits. The dataset I found on Kaggle was good enough for me to explore.
The investment pattern is somehow explored through this EDA. The progress of each of the businesses is yet to be studied.